Split PHCurveLibrary into separate files
